requirements.txt
ultralytics>=8.0.0
roboflow
pyyaml
matplotlib
ipython
